add pagination component
