+++ src/modules/CatalogPage/CatalogPage.module.scss
@@ -70,55 +70,3 @@
     height: 506px;
   }
 }
-
-.circleWithChevrone {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  color: $color__primary;
-  border: 1px solid $color__icons;
-  background-color: $color__white;
-  margin-inline: 12px;
-
-  &:hover {
-    border-color: $color__primary;
-  }
-
-  &:disabled {
-    border-color: $color__elements;
-
-    &> .arrow {
-      filter: none;
-    }
-  }
-}
-
-.circleButton {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  color: $color__primary;
-  border: 1px solid $color__elements;
-  background-color: $color__white;
-  margin-inline: 4px;
-  font-size: 14px;
-  font-weight: 500;
-
-  &:hover {
-    border-color: $color__primary;
-  }
-}
-
-.activeButton {
-  background-color: $color__primary;
-  border: 0;
-  color: $color__white;
-}
-
-.arrow {
-  filter: invert(4%) sepia(45%) saturate(25%) hue-rotate(202deg) brightness(0%)
-    contrast(95%);
-}
