--- src/modules/shared/components/Pagination/index.ts
@@ -1,1 +1,0 @@
-// import * from './Pagination';
